refactor: renamed helper.dart to notification_helper.dart
